Add files via upload
Starting from the 4-color theorem, a modification of the theorem can be made in order to perform triangulations of graphs using only 3 colors. 
This theorem, unlike the 4-color theorem, can be applied to closed figures by means of a convex shell, which delimits the region of the figure. 
The vertices will be joined, each one of them will have an adjacency list where the vertices are adjacent to it, when coloring it will be checked that the vertices are of different colors, once checked that they have a different color, a new edge will be added between the two vertices of different color.